Fix timeout detection and reporting
Replace a 15 seconds-per-process timeout by a global 30 seconds
timeout.
In the previous code, if the first process timed out, then we would
not report it and wait again for the second process. If the second
process terminated within the timeout, and the first process had
exited before, the test would pass. But if the second process
finished first, then the test would fail.
Also, in case of a timeout, the processes were not terminated
which could cause resources issues for the next test (socket already
in use).